+++ formio_components_synchronizer/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 1.0
+
+- Add fields to the Component record, to show the (full, input) paths with keys and labels in the related Form Builder.
+- Distinguish input and layout components. Therefor add `formio.component` boolean field `input`.
+- Other improvements.
+
 ## 0.2
 
 - Fix `name_get` method.